pyproject.toml
[project]
name = "dbt-weather"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dbt-core>=1.11.2",
    "dbt-postgres>=1.10.0",
    "harlequin>=2.1.2",
    "harlequin-postgres>=1.3.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
